requirements.txt
django==6.0.3
psycopg2-binary==2.9.11
pillow==12.2.0
whitenoise==6.12.0
gunicorn
dj-database-url